Cache basic authentications
In order to improve the response time with basic authentication, cache
the results of bcrypt in memory.

This is a draft to see if that approach would be acceptable.

The code is coming from Caddy, and I have fullfiled the attribution
required by Apache-2 by adding the copyright to our headers:
https://github.com/caddyserver/caddy/commit/9a7756c6e4b4ddb945bede3ddb2dfbf241208915

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>