Limit basicauth cache size
If you ended up with 1000 users in your web config, we would have
pre-allocated a map with 1000 entries. The maximum size of the map is
currently hardcoded to 100, so 900 of the preallocations would never be
used.

This limits the initial pre-allocation to 10% of the map size, which is
at this moment 10.

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>