HTTP Headers support
This pull request let users configure the HTTP headers as recommended by
the 2020 Cure53 security audit on Node Exporter.

It is not an arbitrary way to inject any HTTP headers as some backends
(e.g. Prometheus) already add headers (CORS), and we do not want to let
users change those.

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>