Users: Add request logging
This patch adds new configuration options to the web.config.file which
makes it possible to enable logging of http requests.

The "request_log_config" struct takes two arguments, a "file" which
gets written to on request and "header_for_ip" which optionally
allows a user to override how the ip gets retrieved.

A log entry is formatted as json and currently contains the ip
of the request, if authentication was successful, the url path
and a timestamp

Signed-off-by: networkException <git@nwex.de>