Change format of listen on vsock
Before this change, the expected format of listening on
vsock was "--web.listen-address=vsock://:9100".
However, this URI-inspired format is probably confusing,
especially given that it doesn't really support the host
portion of the URI and that listening on TCP doesn't
support the "http://:9100" syntax.
This commit changes the expected format for vsock to be
"--web.listen-address=vsock:9100".

Signed-off-by: Leonid Podolny <leonid@podolny.net>