basic auth: add option to exclude bcrypt at build time
This add the option to exclude bcrypt during the build process by
passing `-tags nobcrypt`. Currently this disables user authentication
via basic_auth. If authorized users are configured but the server is
build with `nobcrypt` the endpoint will respond with Unauthorized, even
if the correct credentials are send.
The goal is to be able to exclude bcrypt for compliance reasons. In the
future we could add basic_auth using other hash functions.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>