Allow `web.listen-address` to be unix domain sockets
 ## Context / Background
Prometheus push gateway is used extensively in the systems I work on. The
current setup I use is to have the pushgateway container running adjacent
to my application and to push metrics over the TCP loopback interface. As
the usage of the pushgateway has scaled up, the overhead and reliability
have become worse.

Using unix domain sockets would allow metrics to be collected in a much
quicker and reliable way.

 ## In this PR
- Added an ability to set `--web.listen-address` to `unix:/path/to/sock`
as a way of specifying that the server should open a listener using
"unix" domain sockets.

 ## How this was tested
- I have made patches against prometheus push gateway and made a
proof-of-concept that seems to work correctly.

Signed-off-by: nick comer <nicholas@cloudflare.com>